+++ js/src/test/scala/com/raquo/waypoint/PageSpec.scala
@@ -8,11 +8,11 @@ class PageSpec extends UnitSpec {
 
   it("read write") {
 
-    write(HomePage) shouldBe "{\"$type\":\"com.raquo.waypoint.fixtures.AppPage.HomePage\"}"
+    write(HomePage) shouldBe "\"com.raquo.waypoint.fixtures.AppPage.HomePage\""
 
     write[AppPage](TextPage("abc123")) shouldBe "{\"$type\":\"com.raquo.waypoint.fixtures.AppPage.TextPage\",\"text\":\"abc123\"}"
 
-    write[AppPage](HomePage)(AppPage.rw) shouldBe "{\"$type\":\"com.raquo.waypoint.fixtures.AppPage.HomePage\"}"
+    write[AppPage](HomePage)(AppPage.rw) shouldBe "\"com.raquo.waypoint.fixtures.AppPage.HomePage\""
 
   }
 }